tec(domain,db): widen rate-limit storage to f64
Switch RateLimitAmount/RateLimitUsage VOs and the
authentication_keys / subscription_types rate-limit columns from i32 to
DOUBLE PRECISION so subscriptions can express fractional quotas. v1
auth envelope stays byte-identical: rate_limit_usage stays i32 input,
rate_limit_remaining is computed as remaining.ceil() as i32 on the
edge.

- New migration 006 promotes rate_limit_daily/remaining/amount.
- Postgres adapter rewrites bound casts from ::int to ::double precision
  and decodes f64 from RETURNING clauses.
- Drops Eq/Hash/Ord derives on the rate-limit VOs and downstream types
  that transitively contained them (RateLimitLedger, IssuedKey,
  AuthDecision, ConsumeOutcome); IEEE-754 forbids them on f64.
- Adds InvalidReason::NotFinite for NaN/Inf rejection in the VO
  constructors.
- Admin DTOs / UI forms accept fractional input (step="any").
- Tests updated to use f64 literals; new pure test pins fractional
  arithmetic.

+++ src/models.rs
@@ -8,8 +8,8 @@ pub struct AuthenticationKey {
     pub key: String,
     pub device_id: String,
     pub subscription: Option<String>,
-    pub rate_limit_daily: i32,
-    pub rate_limit_remaining: i32,
+    pub rate_limit_daily: f64,
+    pub rate_limit_remaining: f64,
     pub rate_limit_updated_at: Option<DateTime<Utc>>,
     pub username: Option<String>,
     pub email: Option<String>,
@@ -42,7 +42,7 @@ pub struct SubscriptionType {
     pub id: i32,
     pub name: String,
     pub display_name: String,
-    pub rate_limit_amount: i32,
+    pub rate_limit_amount: f64,
     pub rate_limit_interval_id: i32,
     pub is_active: bool,
     pub created_at: DateTime<Utc>,
@@ -69,11 +69,11 @@ impl AuthenticationKey {
     }
 
     pub fn rate_limit_percentage(&self) -> f64 {
-        if self.rate_limit_daily == 0 {
+        if self.rate_limit_daily == 0.0 {
             return 0.0;
         }
         let used = self.rate_limit_daily - self.rate_limit_remaining;
-        (used as f64 / self.rate_limit_daily as f64) * 100.0
+        (used / self.rate_limit_daily) * 100.0
     }
 
     pub fn masked_key(&self) -> String {
@@ -93,7 +93,7 @@ pub struct CreateKeyRequest {
     pub email: Option<String>,
     pub subscription: Option<String>,
     pub subscription_type_id: Option<i32>,
-    pub rate_limit_daily: Option<i32>,
+    pub rate_limit_daily: Option<f64>,
     pub expired_at: Option<String>,
 }
 
@@ -104,7 +104,7 @@ pub struct UpdateKeyRequest {
     pub email: Option<String>,
     pub subscription: Option<String>,
     pub subscription_type_id: Option<i32>,
-    pub rate_limit_daily: Option<i32>,
+    pub rate_limit_daily: Option<f64>,
     pub expired_at: Option<String>,
 }
 
@@ -112,7 +112,7 @@ pub struct UpdateKeyRequest {
 pub struct CreateSubscriptionTypeRequest {
     pub name: String,
     pub display_name: String,
-    pub rate_limit_amount: i32,
+    pub rate_limit_amount: f64,
     pub rate_limit_interval_id: i32,
 }
 
@@ -120,7 +120,7 @@ pub struct CreateSubscriptionTypeRequest {
 pub struct UpdateSubscriptionTypeRequest {
     pub name: Option<String>,
     pub display_name: Option<String>,
-    pub rate_limit_amount: Option<i32>,
+    pub rate_limit_amount: Option<f64>,
     pub rate_limit_interval_id: Option<i32>,
     pub is_active: Option<bool>,
 }
